Parse the rates file.
Use the built in ruby XML parser as the needs for this class are simple.
This class is not as dynamic as the transaction parser, as the
requirement for rates is not foreseen to change in the future; where as
transactions may change attribute fields.

It was considered to have a class dedicated to the rate, (an
`ExchangeRate`), however, at this time the rate is simple enough to use
an `OpenStruct`. This does open the possibility of mutability, but at
this time `OpenStruct` is sufficient using freeze.