Adjust exchange algorithm to allow indirect routes
It's possible that different exchanges have different methods for
determining which indirect conversion to use. Some may offer a better
rate that others. To accommodate this allow for the conversion path
mapper to be passed into the exchange. Default to a best conversion
algorithm.

The best conversion algorithm is not complete. Right now it's just a
first found algorithm.